Small updates to existing functions and added a few new one. Added code in main program to read data from sensors periodically. That code also update advertising payload if data was read properly.
Renamed functions:
1) bl_init_buffers() to bl_buffers_init()
2) bl_advertON to bl_advertUpdate()

New functions:
1) bl_buffers_reserve() - to reserve one of bluetooth UART buffer
2) bl_buffers_release() - (force) set all bluetooth UART buffers as free to use
3) bl_buffer_clearStatus() - restore bluetooth UART structure ( buffers, semaphore, etc ) as default
4) bl_advertisingON() - to set bluetooth module in advertising mode

+++ Src/mpl3115_lib.c
@@ -125,7 +125,7 @@ uint32_t mpl_ReadStore_Data(
 		// Convert and store pressure at sea level:
 		((BL_Data_TypeDef *)p_storage)->local_data.press_sea = mpl_get_Psea( p_buff ) ;
 		// Convert and store temperature:
-		((BL_Data_TypeDef *)p_storage)->local_data.temp_tot = p_buff[3] ;
+		((BL_Data_TypeDef *)p_storage)->local_data.temp_tot = (uint8_t)p_buff[3] ;
 		((BL_Data_TypeDef *)p_storage)->local_data.temp_frac = mpl_get_Tempfrac( p_buff[4] );
 	
 		return 1 ;
